fix: Change Image.sizeBytes from Int to BigInt to handle large Docker images
Fixes #90 - Integer overflow when storing Docker images larger than ~2.14GB

Changes:
- Changed Image.sizeBytes field from Int? to BigInt? in Prisma schema
- Updated DatabaseAdapter to use BigInt() for size conversions
- Fixed API routes to properly serialize BigInt values as strings for JSON
- Fixed frontend components to convert BigInt strings back to numbers for calculations
- Generated and applied database migration to alter column type to BIGINT

The fix now supports Docker images up to 9.2 exabytes in size.

+++ src/app/api/scans/local-bulk/route.ts
@@ -67,7 +67,7 @@ export async function POST(request: NextRequest) {
               tag: image.tag,
               source: 'LOCAL_DOCKER',
               platform: 'linux/amd64',
-              sizeBytes: parseInt(image.size.replace(/[^\d]/g, '')) * 1024 * 1024, // Convert MB to bytes
+              sizeBytes: BigInt(parseInt(image.size.replace(/[^\d]/g, '')) * 1024 * 1024), // Convert MB to bytes
               updatedAt: new Date()
             },
             create: {
@@ -76,7 +76,7 @@ export async function POST(request: NextRequest) {
               digest: imageDigest,
               source: 'LOCAL_DOCKER',
               platform: 'linux/amd64',
-              sizeBytes: parseInt(image.size.replace(/[^\d]/g, '')) * 1024 * 1024, // Convert MB to bytes
+              sizeBytes: BigInt(parseInt(image.size.replace(/[^\d]/g, '')) * 1024 * 1024), // Convert MB to bytes
             }
           });
